Move dummy data generator into scripts
