+++ scripts/__init__.py
@@ -1,0 +1,1 @@
+"""Local development and maintenance scripts."""
